Add benchmark to checkout mutations
